// src/shared/api/models/postApiProductBody.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * API
 * OpenAPI spec version: 1.0.0
 */
import type { PostApiProductBodyCategoryId } from './postApiProductBodyCategoryId';
import type { PostApiProductBodySupplierId } from './postApiProductBodySupplierId';
import type { PostApiProductBodyStock } from './postApiProductBodyStock';

export type PostApiProductBody = {
    /** @minLength 1 */
    name: string;
    description?: string;
    image?: string;
    sku?: string;
    categoryId?: PostApiProductBodyCategoryId;
    supplierId?: PostApiProductBodySupplierId;
    brand?: string;
    price?: number;
    costPrice?: number;
    isActive?: boolean;
    availableOnline?: boolean;
    trackInventory: boolean;
    stock?: PostApiProductBodyStock;
    taxIds?: string[];
    lowStockAlert?: boolean;
    lowStockAlertQuantity?: number;
};