// src/shared/api/models/getApiListSuppliers200Item.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * API
 * OpenAPI spec version: 1.0.0
 */
import type { GetApiListSuppliers200ItemAddress } from './getApiListSuppliers200ItemAddress';
import type { GetApiListSuppliers200ItemCity } from './getApiListSuppliers200ItemCity';
import type { GetApiListSuppliers200ItemZipCode } from './getApiListSuppliers200ItemZipCode';
import type { GetApiListSuppliers200ItemCountry } from './getApiListSuppliers200ItemCountry';
import type { GetApiListSuppliers200ItemEmail } from './getApiListSuppliers200ItemEmail';
import type { GetApiListSuppliers200ItemCountryCode } from './getApiListSuppliers200ItemCountryCode';
import type { GetApiListSuppliers200ItemPhone } from './getApiListSuppliers200ItemPhone';
import type { GetApiListSuppliers200ItemCountryISOCode } from './getApiListSuppliers200ItemCountryISOCode';
import type { GetApiListSuppliers200ItemWebsite } from './getApiListSuppliers200ItemWebsite';
import type { GetApiListSuppliers200ItemDescription } from './getApiListSuppliers200ItemDescription';
import type { GetApiListSuppliers200ItemCvrNumber } from './getApiListSuppliers200ItemCvrNumber';
import type { GetApiListSuppliers200ItemContactPersonName } from './getApiListSuppliers200ItemContactPersonName';
import type { GetApiListSuppliers200ItemContactPersonEmail } from './getApiListSuppliers200ItemContactPersonEmail';
import type { GetApiListSuppliers200ItemContactPersonPhone } from './getApiListSuppliers200ItemContactPersonPhone';
import type { GetApiListSuppliers200ItemOutletId } from './getApiListSuppliers200ItemOutletId';

export type GetApiListSuppliers200Item = {
    id: string;
    /**
     * @minimum -2147483648
     * @maximum 2147483647
     */
    seqId: number;
    name: string;
    address: GetApiListSuppliers200ItemAddress;
    city: GetApiListSuppliers200ItemCity;
    zipCode: GetApiListSuppliers200ItemZipCode;
    country: GetApiListSuppliers200ItemCountry;
    email: GetApiListSuppliers200ItemEmail;
    countryCode: GetApiListSuppliers200ItemCountryCode;
    phone: GetApiListSuppliers200ItemPhone;
    countryISOCode: GetApiListSuppliers200ItemCountryISOCode;
    website: GetApiListSuppliers200ItemWebsite;
    description: GetApiListSuppliers200ItemDescription;
    cvrNumber: GetApiListSuppliers200ItemCvrNumber;
    contactPersonName: GetApiListSuppliers200ItemContactPersonName;
    contactPersonEmail: GetApiListSuppliers200ItemContactPersonEmail;
    contactPersonPhone: GetApiListSuppliers200ItemContactPersonPhone;
    outletId: GetApiListSuppliers200ItemOutletId;
    createdAt: string;
    updatedAt: string;
};